rsched: improve grouped CPU metrics
When we do metrics aggregated across all the CPUs, the output is a
little confusing.  We bucket the CPUs by latency, but it's hard to tell
how many CPUs are in each bucket.  It's also hard to see what the global
numbers are.

This commit adds a CPU count to each bucket, and then also histograms
globally.

Signed-off-by: Chris Mason <clm@fb.com>